requirements.txt
numpy==1.19.3
opencv-python==4.4.0.46
pandas==1.1.4
Pillow==8.0.1
pytesseract==0.3.6
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
tesseract==0.1.3
